c_int -> i64

+++ src/codec/encoder/encoder.rs
@@ -70,13 +70,13 @@ impl Encoder {
 
 	pub fn set_bit_rate(&mut self, value: usize) {
 		unsafe {
-			(*self.as_mut_ptr()).bit_rate = value as c_int;
+			(*self.as_mut_ptr()).bit_rate = value as i64;
 		}
 	}
 
 	pub fn set_max_bit_rate(&mut self, value: usize) {
 		unsafe {
-			(*self.as_mut_ptr()).rc_max_rate = value as c_int;
+			(*self.as_mut_ptr()).rc_max_rate = value as i64;
 		}
 	}
